requirements.txt
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.5
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dotenv==1.0.0
